entregasalpes/terceros/requirements.txt
asgiref==3.6.0
attrs==22.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
coverage==6.5.0
cryptography==39.0.2
fastavro==0.24.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
flask-swagger==0.2.14
greenlet==2.0.1
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.32
packaging==21.3
pluggy==1.0.0
protobuf==3.20.3
pulsar-client==2.10.2
py==1.11.0
pycparser==2.21
PyDispatcher==2.0.6
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.44
tomli==2.0.1
urllib3==1.26.14
Werkzeug==2.2.2
